In tools/corcompare: * Makefile (PROGRAM_INSTALL_DIR): Change to $(prefix)/lib/mono/$(FRAMEWORK_VERSION).
In tools/security:
* Makefile (PROGRAM_INSTALL_DIR): Define using $(FRAMEWORK_VERSION).

In mcs:
* Makefile (PROGRAM_INSTALL_DIR): Remove.

In gmcs:
* Makefile: Convert to use executable.make.
* gmcs.exe.sources: New.

In build:
* library.make ($(the_lib)): Don't complain if a .mdb file wasn't
created (even though it's just a warning).
(install-local) [LIBRARY_INSTALL_DIR]: Likewise.
(PACKAGE): Remove.
(install-local) [!LIBRARY_INSTALL_DIR]: Use $(FRAMEWORK_VERSION).
* executable.make ($(PROGRAM)): Likewise.
(install-local): Likewise.
(PROGRAM_INSTALL_DIR): Define Using $(FRAMEWORK_VERSION).

In errors:
* Makefile (DISTFILES): Add do-tests.pl.

svn path=/trunk/mcs/; revision=36628

+++ mcs/build/executable.make
@@ -31,14 +31,13 @@ install-local uninstall-local:
 else
 
 ifndef PROGRAM_INSTALL_DIR
-PROGRAM_INSTALL_DIR = $(prefix)/bin
+PROGRAM_INSTALL_DIR = $(prefix)/lib/mono/$(FRAMEWORK_VERSION)
 endif
-
 
 install-local: $(PROGRAM) $(PROGRAM_config)
 	$(MKINSTALLDIRS) $(DESTDIR)$(PROGRAM_INSTALL_DIR)
 	$(INSTALL_BIN) $(PROGRAM) $(DESTDIR)$(PROGRAM_INSTALL_DIR)
-	-$(INSTALL_BIN) $(PROGRAM).mdb $(DESTDIR)$(PROGRAM_INSTALL_DIR)
+	test ! -f $(PROGRAM).mdb || $(INSTALL_BIN) $(PROGRAM).mdb $(DESTDIR)$(PROGRAM_INSTALL_DIR)
 ifdef PROGRAM_config
 	$(INSTALL_DATA) $(PROGRAM_config) $(DESTDIR)$(PROGRAM_INSTALL_DIR)
 endif
@@ -73,7 +72,7 @@ $(PROGRAM): $(BUILT_SOURCES) $(EXTRA_SOURCES) $(response)
 	$(PROGRAM_COMPILE) /target:exe /out:$(base_prog) $(BUILT_SOURCES) $(EXTRA_SOURCES) @$(response)
 ifneq ($(base_prog),$(PROGRAM))
 	mv $(base_prog) $(PROGRAM)
-	-mv $(base_prog).mdb $(PROGRAM).mdb
+	test ! -f $(base_prog).mdb || mv $(base_prog).mdb $(PROGRAM).mdb
 endif
 
 ifdef PROGRAM_config
